Add support for speed zones

+++ src/repositories/ZoneRepository.js
@@ -44,7 +44,7 @@ class ZoneRepository {
     async getZones() {
         const zoneResult = await this.sql`
             WITH gps_zones AS (
-                SELECT id, type, name, ST_Transform(geom, 4326)
+                SELECT id, type, effect_value, name, ST_Transform(geom, 4326)
                 FROM zones
             )
             SELECT json_build_object(
@@ -57,11 +57,18 @@ class ZoneRepository {
     }
 
     async createZone(zone) {
+        let effectValue = zone.properties.effectValue
+
+        if (effectValue === undefined) {
+            effectValue = null
+        }
+
         const result = await this.sql`
-            INSERT INTO zones (type, name, geom)
+            INSERT INTO zones (type, name, effect_value, geom)
             VALUES (
                 ${zone.properties.type},
                 ${zone.properties.name},
+                ${effectValue},
                 ST_Transform(St_GeomFromGeoJSON(${JSON.stringify(zone.geometry)}), 3857)
             )
             RETURNING id;
@@ -69,44 +76,36 @@ class ZoneRepository {
         return result[0].id
     }
 
-    async deleteZone(id) {
-        await this.sql`
-            DELETE FROM zones WHERE id=${id}
-        `
-    }
-
     async deleteZones(ids) {
+        if (!ids) return
+
         await this.sql`
             DELETE FROM zones WHERE id IN ${ this.sql(ids) }
         `
     }
 
-    async getOverlappingPaths(zoneIds) {
+    async getPathsOverlappingZones() {
         const result = await this.sql`
             WITH unnested_nodes AS (
                 SELECT
-                    intersections.zone_id,
                     ways.id AS way_id,
+                    ways.tags[8] as speed,
                     unnest(array(SELECT nodes[i] FROM generate_series(array_lower(nodes, 1), array_upper(nodes, 1)) i)) AS node_id,
                     generate_series(array_lower(ways.nodes, 1), array_upper(ways.nodes, 1)) AS node_pos
                 FROM
                     (
                         SELECT
-                            zones.id AS zone_id,
-                            osm_id,
-                            (ST_Dump(ST_Intersection(lines.way, zones.geom))).geom AS clip
+                            osm_id
                         FROM planet_osm_line AS lines, zones
-                        WHERE zones.id = ANY(${zoneIds}::int[])
+                        WHERE ST_Intersects(lines.way, zones.geom)
                     ) AS intersections
                 INNER JOIN
                     planet_osm_ways AS ways ON intersections.osm_id = ways.id
-                WHERE
-                    ST_Dimension(intersections.clip) = 1
             ),
             located_nodes AS (
                 SELECT DISTINCT
-                    zone_id,
                     way_id,
+                    speed,
                     node_pos,
                     node_id,
                     n.lat,
@@ -117,18 +116,18 @@ class ZoneRepository {
                     planet_osm_nodes AS n ON node_id = n.id
             )
             SELECT
-                zone_id,
                 way_id,
+                ANY_VALUE(speed) AS speed,
                 ARRAY_AGG(ARRAY[node_id, lat, lon] ORDER BY node_pos) AS nodes
             FROM
                 located_nodes
             GROUP BY
-                zone_id, way_id;
+                way_id;
         `
 
         return result.map(
             row => ({
-                zoneId: row.zone_id,
+                speed: row.speed,
                 nodes: (() => {
                     const path = new Map()
 
@@ -140,53 +139,6 @@ class ZoneRepository {
                 })()
             })
         )
-    }
-
-    async getAllZones() {
-        return await this.sql`
-            SELECT id, ARRAY_AGG(ARRAY[ST_X(dp), ST_Y(dp)]) points
-            FROM (
-                SELECT id, ST_Transform((ST_DumpPoints(geom)).geom, 4326) dp
-                FROM zones
-            )
-            GROUP BY id;
-        `
-    }
-
-    async getAllZonesAndOverlappingPaths() {
-        const pathsResult = await this.sql`
-            SELECT w.id way_id, ARRAY_AGG(n.id) node_ids,
-                ARRAY_AGG(n.lat) node_latitudes,
-                ARRAY_AGG(n.lon) node_longitudes
-            FROM planet_osm_nodes AS n, (
-                    SELECT osm_id, (ST_Dump(ST_Intersection(l.way, z.geom))).geom clip
-                    FROM planet_osm_line AS l, zones AS z
-                ) AS q
-            INNER JOIN planet_osm_ways AS w ON q.osm_id=w.id
-            WHERE ST_Dimension(q.clip)=1 AND n.id=ANY(w.nodes)
-            GROUP BY w.id;
-        `
-
-        const paths = pathsResult.map(
-            row => row.node_ids.map(
-                (id, i) => ({ id: id, lat: row.node_latitudes[i], lon: row.node_longitudes[i] })
-            )
-        )
-
-        const zonesResult = await this.sql`
-            SELECT ARRAY_AGG(ARRAY[ST_X(dp), ST_Y(dp)]) points
-            FROM (
-                SELECT id, ST_Transform((ST_DumpPoints(geom)).geom, 4326) dp
-                FROM zones
-            )
-            GROUP BY id;
-        `
-
-        const zones = zonesResult.map(
-            row => row.points
-        )
-
-        return { paths: paths, zones: zones }
     }
 }
